% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/another_test.R
\name{another_test_function}
\alias{another_test_function}
\title{Another test function}
\usage{
another_test_function(word = "test")
}
\arguments{
\item{word}{This is word will be printed in the terminal}
}
\description{
This function is a test.
}
\examples{
another_test_function(word)
}
\keyword{another}
\keyword{test}
