% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_func.R
\name{cluster_topic}
\alias{cluster_topic}
\title{cluster_topic}
\usage{
cluster_topic(r)
}
\arguments{
\item{r}{a rlda object}
}
\description{
compute percentage of old documents dominated by a given cluster
}
