% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sim.R
\name{compute_sim}
\alias{compute_sim}
\title{compute_sim}
\usage{
compute_sim(r)
}
\arguments{
\item{r}{a rlda object}
}
\description{
compute similarities between topics generated by new models and topics in original models
}
