% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sim_general.R
\name{compute_sim_general}
\alias{compute_sim_general}
\title{compute_sim_general}
\usage{
compute_sim_general(r)
}
\arguments{
\item{r}{a rlda_general object}
}
\description{
compute similarities between topics generated by new models and topics in original models
}
