% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rlda.R
\name{fit}
\alias{fit}
\title{fit_rlda}
\usage{
fit(r)
}
\arguments{
\item{dtm}{document term matrix}

\item{LDA_u}{lda object(in topicmodels)}

\item{K}{numeric or vector, if numeric, number of k to try, if vector, k's to try}

\item{same_k_estimation, }{integer, number of initial states to try, if 0, do not try with diff state}
}
\description{
methods: LDA wrapper, fit
}
