% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_matrix.R
\name{get_cluster_matrix}
\alias{get_cluster_matrix}
\title{get_cluster_matrix}
\usage{
get_cluster_matrix(r, sim_threshold)
}
\arguments{
\item{r}{a rlda object}

\item{sim_threshold}{similarity threshold}
}
\description{
Compute clustering of topics across different models
}
