% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_matrix_general.R
\name{get_cluster_matrix_general}
\alias{get_cluster_matrix_general}
\title{get_cluster_matrix_general}
\usage{
get_cluster_matrix_general(r, sim_threshold)
}
\arguments{
\item{r}{a rlda_general object}

\item{sim_threshold}{similarity threshold}
}
\description{
Compute clustering of topics across different models
}
