% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_in_doc_general.R
\name{getTopicInDoc_general}
\alias{getTopicInDoc_general}
\title{getTopicInDoc_general}
\usage{
getTopicInDoc_general(r)
}
\arguments{
\item{r}{a rlda_general object}
}
\description{
compute percentage of old documents dominated by a given old topic in the new model
}
