% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_topic_in_alt_plot.R
\name{or_topic_in_alt_plot}
\alias{or_topic_in_alt_plot}
\title{or_topic_in_alt_plot}
\usage{
or_topic_in_alt_plot(r, dir)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save plot}
}
\description{
compute clustering of topics across different models
}
