% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_topic_in_alt_plot_general.R
\name{or_topic_in_alt_plot_general}
\alias{or_topic_in_alt_plot_general}
\title{or_topic_in_alt_plot_general}
\usage{
or_topic_in_alt_plot_general(r, dir)
}
\arguments{
\item{r}{a rlda_general object}

\item{dir}{directory to save plot}
}
\description{
compute clustering of topics across different models
}
