% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_author_topic_cov.R
\name{plot_author_topic_cov}
\alias{plot_author_topic_cov}
\title{plot_author_topic_cov}
\usage{
plot_author_topic_cov(r, dir, cluster_info, doc_covs)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save plot}
}
\description{
compute clustering of topics across different models
}
