% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_proportions.R
\name{plot_cluster_proportion}
\alias{plot_cluster_proportion}
\title{plot_cluster_proportions}
\usage{
plot_cluster_proportion(r, dir, cluster_info, doc_covs)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save plot}
}
\description{
compute clustering of topics across different models
}
