% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_proportions_general.R
\name{plot_cluster_proportion_general}
\alias{plot_cluster_proportion_general}
\title{plot_cluster_proportions_general}
\usage{
plot_cluster_proportion_general(r, dir)
}
\arguments{
\item{r}{a rlda_general object}

\item{dir}{directory to save plot}
}
\description{
Plot showing proportion of documents about each topic-cluster.
}
