% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{plot_sim}
\usage{
plot_sim(r, dir)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save generated plots (no "/" at the end)}
}
\description{
generate visualizations for rlda object
}
