% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topDoc.R
\name{plot_topDoc}
\alias{plot_topDoc}
\title{plot_topDoc}
\usage{
plot_topDoc(r, dir)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save generated plots (no "/" at the end)}
}
\description{
generate visualizations for rlda object
}
