% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visual.R
\name{plot_visual}
\alias{plot_visual}
\title{plot_visual}
\usage{
plot_visual(r, dir)
}
\arguments{
\item{r}{a rlda object}

\item{dir}{directory to save generated plots (no "/" at the end)}
}
\description{
generate visualizations for rlda object
}
