% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visual_gen.R
\name{plot_visual_gen}
\alias{plot_visual_gen}
\title{plot_visual_gen}
\usage{
plot_visual_gen(r, dir)
}
\arguments{
\item{r}{a rlda_general object}

\item{dir}{directory to save generated plots (no "/" at the end)}
}
\description{
generate visualizations for rlda_general
}
