% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlda_c.R
\docType{class}
\name{rlda-class}
\alias{rlda-class}
\title{rlda_c}
\description{
Main class
}
\details{
methods: LDA wrapper, fit
}
\section{Slots}{

\describe{
\item{\code{dtm}}{document term matrix}

\item{\code{lda_u}}{lda object(in topicmodels)}

\item{\code{K}}{numeric or vector, if numeric, number of k to try, if vector, k's to try (will be overwrite with list of k's tried once fit has been run)}

\item{\code{compute_parallel}}{bool, if true, will parallel when fitting lda models; otherwise, sequential for fitting}

\item{\code{same_k_estimation}}{integer, number of initial states to try, if 0, do not try with diff state}

\item{\code{similarity_measure}}{string, similarity measure (so far cosine or hellinger). Default: cosine}

\item{\code{num_of_clusters}}{numeric or vector, number of clusters used when performing spectral clustering}

\item{\code{seed_list}}{seeds tried (exists only when same_k_estimation is true)}

\item{\code{beta_list}}{list of beta matrices in LDA objects of all K tried (ordered same as K)}

\item{\code{gamma_list}}{list of gamma matrices in LDA objects of all K tried (ordered same as K)}

\item{\code{model_topic_mat}}{percentage of documents dominated by the given topic}

\item{\code{similarity_mat}}{maximum similarity (given choice of similarity functions) of a given topic compare to any topics in the original lda model (to give the probability of a user’s topic shows up in a tried model’s resulting topics)}

\item{\code{sim_matrix_list}}{list of similarity matrices that gives us similarity between}

\item{\code{init_states}}{# not used yet, probably gonna delete}

\item{\code{key_features}}{top 10 features of a given topic in each model tried}

\item{\code{topic_dom_perc_list}}{percentage of documents dominated by the given topic out of documents originally dominated by similar topic in the original model}

\item{\code{dominant_topic_cluster_list}}{clusters correponding to dominant topics of each document in each model}

\item{\code{cluster_center_key_words_list}}{(data_frame?) top 10 keywords for each center found by the cluster algorithn (so far only support spectral clustering)}

\item{\code{perc_document_belong_cluster_list}}{percentage of documents belong to a given cluster in a given model}

\item{\code{topic_cluster_assignment}}{cluster number a given topic belongs to d}

\item{\code{top_stability_mat}}{a matrix indicating whether each cluster is present in each alternative model}

\item{\code{doc_by_cluster_and_model}}{a matrix indicating the dominiant cluster of each document according to each topic model}
}}

