% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_general.R
\name{set_up_general}
\alias{set_up_general}
\title{set_up_general}
\usage{
set_up_general(r)
}
\arguments{
\item{r}{a rlda_general object}
}
\description{
compute similarities between topics generated by new models and topics in original models
}
