% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_function}
\alias{test_function}
\title{Test function}
\usage{
test_function(word = "test")
}
\arguments{
\item{word}{This is word will be printed in the terminal}
}
\description{
This function is a test.
}
\examples{
test_function(word)
}
\keyword{test}
