% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util}
\alias{util}
\title{util
rlda utility functions
functions: computeSimilarity, saveSimilarity (???), saveClustering, saveTopKws}
\usage{
util()
}
\description{
util
rlda utility functions
functions: computeSimilarity, saveSimilarity (???), saveClustering, saveTopKws
}
