% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_res.R
\name{visual_res}
\alias{visual_res}
\title{visualize_res}
\usage{
visual_res(r)
}
\arguments{
\item{r}{a rlda object}
}
\description{
compute similarities between topics generated by new models and topics in original models
}
