% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rlda.R
\name{fit}
\alias{fit}
\title{fit_rlda}
\usage{
fit(r)
}
\arguments{
\item{r}{a rlda or rstm object}
}
\description{
This function fits different LDA models with different number of clusters for a rlda object
}
