% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_in_doc.R
\name{getTopicInDoc}
\alias{getTopicInDoc}
\title{getTopicInDoc}
\usage{
getTopicInDoc(r, sim_threshold)
}
\arguments{
\item{r}{a rlda object}

\item{sim_threshold}{similarity threshold}
}
\description{
compute percentage of old documents dominated by a given old topic in the new model
}
