% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_in_doc.R
\name{getTopicInDoc}
\alias{getTopicInDoc}
\title{getTopicInDoc}
\usage{
getTopicInDoc(r)
}
\arguments{
\item{r}{a rlda object}
}
\description{
compute percentage of documents dominated by a given topic
}
