% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_author_topic_cov.R
\name{plot_author_topic_cov}
\alias{plot_author_topic_cov}
\title{plot_author_topic_cov}
\usage{
plot_author_topic_cov(r, dir, doc_covs, covs)
}
\arguments{
\item{r}{An rlda object.}

\item{dir}{Directory to save the resulting plot.}

\item{doc_covs}{A dataset with author and/or document-level covariates. The dataset needs to be indexed exactly as the documents in the Document Term Matrix provided during fitting.}

\item{covs}{A vector with the covariates to analayze: 5 covariates maximum.}
}
\description{
This function plots the bivariate relationship between a set of author and/or document-level covariates and the likelihood of a document to be about each of the topic clusters.
}
